% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_package.R
\name{plssemMI-package}
\alias{plssemMI-package}
\alias{plssemMI-pkg}
\alias{plssemMI-p}
\alias{plssemMI}
\title{plssemMI: a package}
\description{
The \pkg{plssemMI} package contains functions that implements
different missing data imputation approaches for partial least
squares structural equation models (PLS-SEM).

The main functions included in the package are:
\itemize{
\item \code{\link{make_missing}()}
}
}
\details{
_PACKAGE
}
\keyword{internal}
