\name{plssemMI-package}
\alias{plssemMI-package}
\alias{plssemMI}
\docType{package}
\title{
\packageTitle{plssemMI}
}
\description{
\packageDescription{plssemMI}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{plssemMI}

Maintainer: \packageMaintainer{plssemMI}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = plssemMI)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{plssemMI}}
%\section{Documentation Index}{\packageIndices{plssemMI}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
